Homework module 46 1. Dependencies versions and repeatable dependencies were moved to versions.gradle. 2. Retrofit was moved to separate module.
