+++ retrofit/src/main/java/film/search/retrofit/RemoteComponent.kt
@@ -1,0 +1,10 @@
+package film.search.retrofit
+
+import dagger.Component
+import javax.inject.Singleton
+
+@Singleton
+@Component(
+    modules = [RemoteModule::class]
+)
+interface RemoteComponent : RemoteProvider
