+++ retrofit/src/main/java/film/search/retrofit/RemoteProvider.kt
@@ -1,0 +1,5 @@
+package film.search.retrofit
+
+interface RemoteProvider {
+    fun provideRemote(): TmdbApi
+}
