+++ settings.gradle
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "FilmsSearch"
 include ':app'
+include ':retrofit'
