+++ versions.gradle
@@ -1,0 +1,15 @@
+ext {
+    room_version = '2.6.1'
+    okhttp_version = '4.12.0'
+    retrofit_version = '2.9.0'
+    dagger_version = '2.50'
+    coroutines_version = '1.7.3'
+    glide_version = '4.16.0'
+
+    dagger = "com.google.dagger:dagger:$dagger_version"
+    daggerCompiler = "com.google.dagger:dagger-compiler:$dagger_version"
+    roomRuntime = "androidx.room:room-runtime:$room_version"
+    roomCompiler = "androidx.room:room-compiler:$room_version"
+    rxandroid = 'io.reactivex.rxjava3:rxandroid:3.0.2'
+    rxjava = 'io.reactivex.rxjava3:rxjava:3.1.8'
+}
